Rename the module with `Orientation` struct to `metadata`
